# web-ui/backend/openapi.yaml
openapi: 3.0.3
info:
  title: AI PowerShell Assistant Web API
  description: RESTful API for AI PowerShell Assistant Web UI
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/templates:
    get:
      tags:
        - Template
      summary: Get template list
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Templates retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Template'

    post:
      tags:
        - Template
      summary: Create new template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreate'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          description: Invalid template data
  /api/templates/{id}:
    get:
      tags:
        - Template
      summary: Get template details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          description: Template not found

    put:
      tags:
        - Template
      summary: Update template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateUpdate'
      responses:
        '200':
          description: Template updated
        '404':
          description: Template not found

    delete:
      tags:
        - Template
      summary: Delete template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template deleted
        '404':
          description: Template not found
  /api/templates/{id}/generate:
    post:
      tags:
        - Template
      summary: Generate script from template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                parameters:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: Script generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      script:
                        type: string
...
components:
  schemas:
...
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        scriptContent:
          type: string
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/TemplateParameter'
        keywords:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TemplateParameter:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [string, number, boolean, select]
        required:
          type: boolean
        default:
          nullable: true
        options:
          type: array
          items:
            type: string
        description:
          type: string
    TemplateCreate:
      type: object
      required:
        - name
        - category
        - scriptContent
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        scriptContent:
          type: string
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/TemplateParameter'
        keywords:
          type: array
          items:
            type: string
    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        scriptContent:
          type: string
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/TemplateParameter'
        keywords:
          type: array
          items:
            type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT